I just put everything in a class because it was way less complicated
- Added fileManagement as a folder where I have my functions about opening files and having a list of all files, stuff like that. Not in a class.
- My the stitching functions in a class, way less complicated to have self attributes to an object in my case.
Bye the way, its name is now stitching instead of functions

That's all! now I have an error to solve : TypeError: Stitching.__init__() got multiple values for argument 'sourceDir'
