+++ main.py
@@ -1,28 +1,13 @@
-from functions import *
+from stitching import *
 
 sourcePath = "/Users/valeriepineaunoel/Desktop/test"
-files = listNameOfFiles(directory=sourcePath)
 tileDimensions = [2, 3]
 # [x,y]. Positive y = going down. Positive x = going right. When comparing image2 with image1 : 
 verticalShift = [39, 384]
 horizontalShift = [169, -12]
 
-# initialize images in numpy arrays
-imagenp0 = read_file(filePath=sourcePath + "/" + files[0], imageType="numpy")
-imagenp1 = read_file(filePath=sourcePath + "/" + files[1], imageType="numpy")
-#image_offset = read_file(filePath=sourcePath + "/" + files[2], imageType="numpy")
-imagenp2 = read_file(filePath=sourcePath + "/" + files[2], imageType="numpy")
-imagenp3 = read_file(filePath=sourcePath + "/" + files[3], imageType="numpy")
+print(sourcePath)
 
-# initialise images as pillow images
-imagePIL0 = read_file(filePath=sourcePath + "/" + files[0], imageType="PIL", mirror=True)
-imagePIL1 = read_file(filePath=sourcePath + "/" + files[1], imageType="PIL", mirror=True)
-imagePIL2 = read_file(filePath=sourcePath + "/" + files[2], imageType="PIL", mirror=True)
-imagePIL3 = read_file(filePath=sourcePath + "/" + files[3], imageType="PIL", mirror=True)
-
-backgroundImage = create_tile_image(tileD=tileDimensions, hShift=horizontalShift, vShift=verticalShift)
-# width, height = backgroundImage.size
-backgroundImage.save(fp="/Users/valeriepineaunoel/Desktop/backgroundImage.tif")
-
-stitchedImage = stitching_scrapbooking(tileD=tileDimensions, image=imagePIL0, tile=backgroundImage, vShift=verticalShift, hShift=horizontalShift)
-stitchedImage.save(fp="/Users/valeriepineaunoel/Desktop/tile.tif")
+stitch = Stitching(sourceDir=sourcePath, tileD=tileDimensions, imageSize=[1024,512], vShift=verticalShift, hShift=horizontalShift)
+stitchedImage = stitch.stitching_scrapbooking_allImages()
+stitchedImage.save(fp="/Users/valeriepineaunoel/Desktop/stitchedImage.tiff")
